Asp.net project done on visual studio installer 2022
In this project, i've added two more properties in the event class. as this was migrated and updated in database in project manager console and then scaffolded, we have to create another migration with Add-Migration ____ (another name for migration) after migration, update database. after this is done, edit all the files in views folder that were scaffolded before. after all the classes are edited, this work is done.

in addition to this, i also used validation in create.cs. which basically allows only asterisks in rating. and search bar is also implemented.